A little refactoring, exposing request body to mock service
