+++ src/hunit/assert/reflect.go
@@ -1,0 +1,14 @@
+package assert
+
+import "reflect"
+
+// Obtain a value's type and kind
+func typeAndKind(v interface{}) (reflect.Type, reflect.Kind) {
+	t := reflect.TypeOf(v)
+	k := t.Kind()
+	for k == reflect.Ptr {
+		t = t.Elem()
+		k = t.Kind()
+	}
+	return t, k
+}
