+++ src/hunit/entity.go
@@ -1,15 +1,11 @@
 package hunit
 
 import (
-	"bytes"
-	"encoding/csv"
-	"encoding/json"
-	"fmt"
-	"io"
 	"strings"
 	"unicode"
 
-	"github.com/instaunit/instaunit/hunit/httputil/mimetype"
+	"github.com/instaunit/instaunit/hunit/assert"
+	"github.com/instaunit/instaunit/hunit/entity"
 	"github.com/instaunit/instaunit/hunit/runtime"
 	"github.com/instaunit/instaunit/hunit/testcase"
 )
@@ -30,7 +26,7 @@ func literalEntitiesEqual(context runtime.Context, contentType string, expected 
 
 	var abytes []byte
 	if abytes, ok = actual.([]byte); !ok {
-		return &AssertionError{expected, actual, "Entities are not equal"}
+		return &assert.AssertionError{expected, actual, "Entities are not equal"}
 	}
 
 	if (context.Options & testcase.OptionEntityTrimTrailingWhitespace) == testcase.OptionEntityTrimTrailingWhitespace {
@@ -41,8 +37,8 @@ func literalEntitiesEqual(context runtime.Context, contentType string, expected 
 		a = abytes
 	}
 
-	if !equalValues(e, a) {
-		return &AssertionError{e, a, "Entities are not equal"}
+	if !assert.EqualValues(e, a) {
+		return &assert.AssertionError{e, a, "Entities are not equal"}
 	} else {
 		return nil
 	}
@@ -51,126 +47,14 @@ func literalEntitiesEqual(context runtime.Context, contentType string, expected 
 // Compare entities for equality
 func semanticEntitiesEqual(context runtime.Context, contentType string, expected []byte, actual interface{}) error {
 
-	e, err := unmarshalEntity(context, contentType, expected)
+	e, err := entity.Unmarshal(contentType, expected)
 	if err != nil {
 		return err
 	}
 
-	if !semanticEqual(e, actual) {
-		return &AssertionError{e, actual, "Entities are not equal"}
+	if !entity.SemanticEqual(e, actual) {
+		return &assert.AssertionError{e, actual, "Entities are not equal"}
 	} else {
 		return nil
 	}
-}
-
-// Unmarshal an entity
-func unmarshalEntity(context runtime.Context, contentType string, entity []byte) (interface{}, error) {
-
-	// trim off the parameters following ';' if we have any
-	if i := strings.Index(contentType, ";"); i > 0 {
-		contentType = contentType[:i]
-	}
-
-	switch contentType {
-	case mimetype.JSON:
-		return unmarshalJSONEntity(context, entity)
-	case mimetype.CSV:
-		return unmarshalCSVEntity(context, entity)
-	default:
-		return nil, fmt.Errorf("Unsupported content type for semantic comparison: %v", contentType)
-	}
-
-}
-
-// Unmarshal a JSON entity
-func unmarshalJSONEntity(context runtime.Context, entity []byte) (interface{}, error) {
-	if entity == nil || len(entity) < 1 {
-		return nil, nil
-	}
-	var value interface{}
-	err := json.Unmarshal(entity, &value)
-	if err != nil {
-		return nil, fmt.Errorf("Invalid JSON entity: %v", err)
-	}
-	return value, nil
-}
-
-// Unmarshal a CSV entity
-func unmarshalCSVEntity(context runtime.Context, entity []byte) (interface{}, error) {
-	if entity == nil || len(entity) < 1 {
-		return nil, nil
-	}
-
-	value := make([]interface{}, 0)
-	var h []string
-
-	r := csv.NewReader(bytes.NewBuffer(entity))
-	for {
-		row, err := r.Read()
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			return nil, fmt.Errorf("Invalid CSV entity: %v (%s)", err, strings.Join(row, ", "))
-		}
-
-		if h == nil {
-			h = row
-			continue
-		}
-
-		m := make(map[string]string)
-		for i, e := range row {
-			m[h[i]] = e
-		}
-		value = append(value, m)
-	}
-
-	return value, nil
-}
-
-// Compare results
-func semanticEqual(expected, actual interface{}) bool {
-	switch a := actual.(type) {
-
-	case map[string]string:
-		e, ok := expected.(map[string]string)
-		if !ok {
-			return false
-		}
-		for k, v := range e {
-			if !semanticEqual(v, a[k]) {
-				return false
-			}
-		}
-
-	case map[string]interface{}:
-		e, ok := expected.(map[string]interface{})
-		if !ok {
-			return false
-		}
-		for k, v := range e {
-			if !semanticEqual(v, a[k]) {
-				return false
-			}
-		}
-
-	case []interface{}:
-		e, ok := expected.([]interface{})
-		if !ok {
-			return false
-		}
-		if len(a) != len(e) {
-			return false
-		}
-		for i, v := range e {
-			if !semanticEqual(v, a[i]) {
-				return false
-			}
-		}
-
-	default:
-		return equalValues(expected, actual)
-
-	}
-	return true
 }
